Create Email

This function creates an email using the blastula package by composing the email body and footer.

create_email(body, footer, username = NULL, name = NULL, code = NULL, ...)

Arguments

body

The body content of the email.

footer

The footer content of the email.

username

(Optional) The username associated with the email.

name

(Optional) The name of the recipient.

code

(Optional) A code or verification token to be included in the email.

...

Additional arguments to be passed to the blastula::compose_email function.

Value

An email object created using the blastula package.

Examples

# Create an email
email <- create_email(
  body = "Hello {name}, your username is {username}.",
  footer = "Please contact support if you have any questions.",
  username = "john123",
  name = "John Doe"
)